backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
facebook-business==21.0.2
anthropic==0.40.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
pydantic[email]==2.10.3
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
reportlab==4.2.5
Pillow==11.0.0
pandas==2.2.3
